requirements.txt
requests~=2.26.0
spotipy~=2.19.0
beautifulsoup4~=4.9.3
python-dotenv~=0.19.0